Many changes such as :
Added a StreamActivity, Adapter, View and Class. Modified the flow of
the app so that when the user opens it and after he logs in (which is
disabled right now, so login works regardless of what the username and
password is, they can even be empty) a list of stream (hard coded for
now) appears. On click, a new activity opens where the stream will play.